backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
httpx>=0.25.0
aiofiles>=23.0.0
Pillow>=10.0.0
